+++ src/sprite.h
@@ -4,23 +4,24 @@
 #include "common.h"
 
 typedef struct s_sprite {
-    SDL_Surface *image;
-    SDL_Rect rect;
-    SDL_Rect clip_rect;
+  SDL_Surface *image;
+  SDL_Rect rect;
+  SDL_Rect clip_rect;
 
-    int current_frame;
-    int frame_count;
-    int start_position;
+  i64 current_frame;
+  i64 frame_count;
+  i8 start_position;
 } sprite;
 
 typedef struct s_sprite_node {
-    sprite *sprite;
-    struct s_sprite_node *previous;
-    struct s_sprite_node *next;
+  sprite *sprite;
+  struct s_sprite_node *previous;
+  struct s_sprite_node *next;
 } sprite_node;
 
-sprite *load_sprite(char *filename, int width, int height, int x, int y, int frame_count);
-void move_sprite(sprite *spr, int x, int y);
+sprite *load_sprite(char *filename, u16 width, u16 height, u16 x, u16 y,
+                    u8 frame_count);
+void move_sprite(sprite *spr, u16 x, u16 y);
 void update_sprite(sprite *spr);
 void teardown_sprite(sprite *spr);
 
@@ -31,4 +32,3 @@ sprite_node *create_node(sprite *spr);
 void stack_remove(sprite_node *node);
 void stack_push(sprite_node *node);
 #endif
-
